added functionality for centiloid porcessing

+++ spm12/regseg.py
@@ -3,6 +3,7 @@ import logging
 import os
 import shutil
 from textwrap import dedent
+import re
 
 import numpy as np
 import scipy.ndimage as ndi
@@ -15,8 +16,31 @@ __author__ = ("Pawel J. Markiewicz", "Casper O. da Costa-Luis")
 log = logging.getLogger(__name__)
 
 
+
+
+
+def move_files(fin, opth):
+    '''
+    Move input file path fin to the output folder opth.
+    '''
+    fdst = os.path.join(opth, os.path.basename(fin))
+    shutil.move(fin, fdst)
+    return fdst
+
+
+
+def glob_re(pttrn, pth):
+    '''
+    glob with regular expressions
+    '''
+    m = re.compile(pttrn)
+    return [os.path.join(pth,f) for f in os.listdir(pth) if m.match(f)]
+
+
+
 def fwhm2sig(fwhm, voxsize=2.0):
     return fwhm / (voxsize * (8 * np.log(2)) ** 0.5)
+
 
 
 def smoothim(fim, fwhm=4, fout=""):
@@ -29,9 +53,8 @@ def smoothim(fim, fwhm=4, fout=""):
     )
 
     if not fout:
-        fout = "{f[0]}{f[1]}_smo{fwhm}{f[2]}".format(
-            f=nii.file_parts(fim), fwhm=str(fwhm).replace(".", "-")
-        )
+        f = nii.file_parts(fim)
+        fout = os.path.join(f[0], '{}_smo{}{}'.format(f[1], str(fwhm).replace(".", "-"), f[2]))
 
     nii.array2nii(
         imsmo,
@@ -48,6 +71,7 @@ def smoothim(fim, fwhm=4, fout=""):
     return {"im": imsmo, "fim": fout, "fwhm": fwhm, "affine": imd["affine"]}
 
 
+# >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>
 def coreg_spm(
     imref,
     imflo,
@@ -68,7 +92,22 @@ def coreg_spm(
     del_uncmpr=True,
     save_arr=True,
     save_txt=True,
+    modify_nii=False,
 ):
+    ''' Rigid body registration using SPM coreg function.
+        Arguments:
+        imref       - reference image
+        imflo       - floating image
+        fwhm_ref    - FWHM of the smoothing kernel for the reference image
+        fwhm_flo    - FWHM of the smoothing kernel for the reference image
+
+        modify_nii  - modifies the affine of the NIfTI file of the floating
+                      image according to the rigid body transformation. 
+    '''
+
+    # > output dictionary
+    out = {}
+
     sep = sep or [4, 2]
     tol = tol or [
         0.0200,
@@ -128,7 +167,12 @@ def coreg_spm(
     if fwhm_flo > 0:
         smodct = smoothim(imflou, fwhm_flo)
         # > delete the previous version (non-smoothed)
-        os.remove(imflou)
+        if not modify_nii:
+            os.remove(imflou)
+        else:
+            # > save the uncompressed and unsmoothed version
+            imflou_ = imflou
+
         imflou = smodct["fim"]
 
         log.info(
@@ -137,7 +181,7 @@ def coreg_spm(
             )
         )
 
-    # run the matlab SPM coregistration
+    # run the MATLAB SPM registration
     import matlab as ml
 
     Mm, xm = eng.coreg_spm_m(
@@ -152,6 +196,12 @@ def coreg_spm(
         visual,
         nargout=2,
     )
+
+    # > modify the affine of the floating image (as usually done in SPM)
+    if modify_nii: 
+        eng.coreg_modify_affine(imflou_, Mm)
+        out['freg'] = imflou_
+
 
     # get the affine matrix
     M = np.array(Mm._data.tolist())
@@ -198,13 +248,13 @@ def coreg_spm(
         faff = os.path.splitext(faff)[0] + ".txt"
         np.savetxt(faff, M)
 
-    return {
-        "affine": M,
-        "faff": faff,
-        "rotations": x[3:],
-        "translations": x[:3],
-        "matlab_eng": eng,
-    }
+    out['affine'] = M
+    out['faff'] = faff
+    out['rotations'] = x[3:]
+    out['translations'] = x[:3]
+    out['matlab_eng'] = eng
+
+    return out
 
 
 def resample_spm(
@@ -320,3 +370,126 @@ def resample_spm(
         )
 
     return fout
+
+
+
+#>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>
+# SEGMENTATION
+#>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>
+def seg_spm(
+        f_mri,
+        spm_path,
+        matlab_eng_name='',
+        outpath=None,
+        store_nat_gm=False,
+        store_nat_wm=False,
+        store_nat_csf=False,
+        store_fwd=False,
+        store_inv=False,
+        visual=False):
+
+    ''' perform normalisation/segmentation using SPM12.
+        Arguments:
+        f_mri   - file path to the T1w MRI file
+        spm_path- SPM path (string)
+        matlab_eng_name - name of the Python engine for Matlab.  
+        outpath - output folder path for the normalisation file output
+        store_nat_* - stores native space segmentation output for either
+                      grey matter, white matter or CSF.
+        sotre_fwd/inv - stores forward/inverse normalisation definitions
+        visual  - shows the Matlab window progress
+    '''
+
+    # > output dictionary
+    out = {}
+    
+    # > get Matlab engine or use the provided one
+    eng = ensure_spm(matlab_eng_name)
+
+    # > run SPM normalisation/segmentation
+    param,invdef,fordef = eng.seg_spm_m(
+        f_mri,
+        str(spm_path),
+        float(store_nat_gm),
+        float(store_nat_wm),
+        float(store_nat_csf),
+        float(store_fwd),
+        float(store_inv),
+        float(visual),
+        nargout=3
+    )
+
+    #------------------------------------------------------------------
+    # > organise the output
+    if outpath is not None:
+        create_dir(outpath)
+
+        out['param']  = move_files(param, outpath)
+        out['invdef'] = move_files(invdef, outpath)
+        out['fordef'] = move_files(fordef, outpath)
+
+        # > go through tissue types and move them to the output folder
+        cs = glob_re(r'c\d*', os.path.dirname(param))
+        for c in cs:
+            nm = os.path.basename(c)[:2]
+            out[nm] = move_files(c, outpath)
+    
+    else:
+        out['param']  = param
+        out['invdef'] = invdef
+        out['fordef'] = fordef
+
+        cs = glob_re(r'c\d*', os.path.dirname(param))
+        for c in cs:
+            nm = os.path.basename(c)[:2]
+            out[nm] = c
+    #------------------------------------------------------------------
+
+    return out
+
+#>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>
+
+
+
+#>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>
+# NORMALISATION WRITE
+#>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>
+def normw_spm(
+        f_def,
+        files4norm,
+        matlab_eng_name='',
+        outpath=None):
+
+    ''' 
+    Write normalisation output to NIfTI files using SPM12.
+    Arguments:
+    f_def   - NIfTI file of definitions for non-rigid normalisation
+    files4norm - list of input NIfTI file paths in format ['file, 1']
+    matlab_eng_name - name of the Python engine for Matlab.
+    outpath - output folder path for the normalisation files
+    '''
+
+    eng = ensure_spm(matlab_eng_name)  # get_matlab
+
+    eng.normw_spm_m(
+        f_def,
+        files4norm
+    )
+
+    # > output list
+    out = []
+
+    #------------------------------------------------------------------
+    # > organise the output
+    if outpath is not None:
+        create_dir(outpath)
+
+        for f in files4norm:
+            fpth = f.split(',')[0]
+            out.append(move_files(os.path.join(os.path.dirname(fpth), 'w'+os.path.basename(fpth)), outpath))
+    else:
+        out.append('w'+os.path.basename(f.split(',')[0]))
+
+    return out
+
+#>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>
